Remove longclaw homepage summary item override by default,
- Adds support for settings LONGCLAW_OVERRIDE_SUMMARY_ITEMS and LONGCLAW_INCLUDE_SUMMARY_ITEMS

Together these settings allow for longclaw to override the homepage summary items with it own or just add its own to the list.
If both a False (default) then it is assumed that the summary items will be set at the project level, allowing for greater control and interoperability with other projects.